:recycle: 이메일 서비스 리팩터링
리팩터링
